:bone: unify the two event types into one message handler, improve code, comments, fmt, clippy
