+++ src/main.rs
@@ -29,9 +29,7 @@ use web::{web_main, WebState};
 mod args;
 mod command_manager;
 mod console;
-mod console_events;
 mod demo;
-mod demo_events;
 mod events;
 mod gamefinder;
 mod io;
@@ -42,17 +40,17 @@ mod player;
 mod player_records;
 mod server;
 mod settings;
+mod sse_events;
 mod state;
 mod steam_api;
 mod web;
 
 use command_manager::{Command, CommandManager, DumbAutoKick};
 use console::{ConsoleLog, ConsoleOutput, ConsoleParser, RawConsoleOutput};
-use console_events::SseBroadcastMessages;
 use demo::{DemoBytes, DemoManager, DemoMessage, DemoWatcher, PrintVotes};
-use demo_events::SseBroadcastVotes;
 use events::{Preferences, Refresh, UserUpdates};
 use new_players::{ExtractNewPlayers, NewPlayers};
+use sse_events::BroadcastableEvent;
 use steam_api::{
     FriendLookupResult, LookupFriends, LookupProfiles, ProfileLookupBatchTick, ProfileLookupResult,
 };
@@ -83,11 +81,11 @@ define_events!(
         DemoMessage,
     },
     Handler {
+        BroadcastableEvent,
+
         CommandManager,
-
         ConsoleParser,
-        SseBroadcastMessages,
-
+        // SseBroadcastMessages,
         ExtractNewPlayers,
 
         LookupProfiles,
@@ -97,8 +95,7 @@ define_events!(
 
         DemoManager,
         PrintVotes,
-        SseBroadcastVotes,
-
+        // SseBroadcastVotes,
         DumbAutoKick,
     },
 );
@@ -221,8 +218,9 @@ fn main() {
                 .add_handler(LookupFriends::new())
                 .add_handler(DumbAutoKick)
                 .add_handler(WebAPIHandler::new())
-                .add_handler(SseBroadcastMessages)
-                .add_handler(SseBroadcastVotes::new());
+                .add_handler(BroadcastableEvent::new());
+                // .add_handler(SseBroadcastMessages)
+                // .add_handler(SseBroadcastVotes::new());
 
             if args.print_votes {
                 event_loop = event_loop.add_handler(PrintVotes::new());
